Create UI login Page
Flutter Login UI
This project is a simple and elegant Login UI built with Flutter, designed to demonstrate how to create a clean and user-friendly interface for mobile applications. The UI features essential elements of a login screen, including input fields for email and password, a login button, and links for account creation.